graph-reducer.ts
