+++ neo4j-to-rdf.js
@@ -6,7 +6,7 @@ const neo4j = require('neo4j-driver')
 
 const prec = require('./prec.js');
 const RDFGraphBuilder = require("./src/prec/graph-builder");
-const graphReducer    = require("./src/prec/graph-reducer.js");
+const { default: graphReducer } = require("./src/prec/graph-reducer");
 
 const { ArgumentParser } = require('argparse');
