+++ package.json
@@ -4,7 +4,7 @@
     "description": "A Property Graph to RDF converter",
     "main": "prec.js",
     "scripts": {
-        "test": "mocha"
+        "test": "npx ts-mocha"
     },
     "repository": {
         "type": "git",
