+++ prec.js
@@ -9,7 +9,7 @@
 const N3            = require('n3');
 
 const RDFGraphBuilder = require("./src/prec/graph-builder");
-const graphReducer    = require("./src/prec/graph-reducer.js");
+const { default: graphReducer } = require("./src/prec/graph-reducer");
 
 const { ArgumentParser } = require('argparse');
